+++ src/compas/datastructures/datastructure.py
@@ -7,3 +7,69 @@ from compas.data import Data
 
 class Datastructure(Data):
     """Base class for all data structures."""
+
+    def transform(self, transformation):
+        """Transforms the data structure.
+
+        Parameters
+        ----------
+        transformation : :class:`Transformation`
+            The transformation used to transform the data structure.
+
+        Returns
+        -------
+        None
+
+        """
+        raise NotImplementedError
+
+    def transformed(self, transformation):
+        """Returns a transformed copy of this data structure.
+
+        Parameters
+        ----------
+        transformation : :class:`Transformation`
+            The transformation used to transform the copy.
+
+        Returns
+        -------
+        Datastructure
+            The transformed copy.
+
+        """
+        datastructure = self.copy()
+        datastructure.transform(transformation)
+        return datastructure
+
+    def transform_numpy(self, transformation):
+        """Transforms the data structure.
+
+        Parameters
+        ----------
+        transformation : :class:`Transformation`
+            The transformation used to transform the data structure.
+
+        Returns
+        -------
+        None
+
+        """
+        raise NotImplementedError
+
+    def transformed_numpy(self, transformation):
+        """Returns a transformed copy of this data structure.
+
+        Parameters
+        ----------
+        transformation : :class:`Transformation`
+            The transformation used to transform the copy.
+
+        Returns
+        -------
+        Datastructure
+            The transformed copy.
+
+        """
+        datastructure = self.copy()
+        datastructure.transform_numpy(transformation)
+        return datastructure