+++ src/compas/datastructures/mesh/operations/weld.py
@@ -2,7 +2,7 @@ from __future__ import print_function
 from __future__ import absolute_import
 from __future__ import division
 
-from compas.topology import adjacency_from_edges
+from compas.topology import vertex_adjacency_from_edges
 from compas.topology import connected_components
 
 from compas.utilities import pairwise
@@ -86,7 +86,7 @@ def mesh_unweld_edges(mesh, edges):
                     )
                 )
 
-        adjacency = adjacency_from_edges(network_edges)
+        adjacency = vertex_adjacency_from_edges(network_edges)
         for key, values in adjacency.items():
             adjacency[key] = {value: None for value in values}
         # include non connected vertices