+++ src/compas/datastructures/network/network.py
@@ -21,7 +21,23 @@ from compas.geometry import midpoint_line
 from compas.geometry import normalize_vector
 from compas.geometry import add_vectors
 from compas.geometry import scale_vector
+from compas.geometry import transform_points
+from compas.topology import astar_shortest_path
+
 from compas.datastructures import Graph
+
+from .operations.split import network_split_edge
+from .operations.join import network_join_edges
+
+from .planarity import network_is_crossed
+from .planarity import network_is_planar
+from .planarity import network_is_planar_embedding
+from .planarity import network_is_xy
+from .planarity import network_count_crossings
+from .planarity import network_find_crossings
+from .planarity import network_embed_in_plane
+from .smoothing import network_smooth_centroid
+from .duality import network_find_cycles
 
 
 class Network(Graph):
@@ -37,6 +53,20 @@ class Network(Graph):
         Additional attributes to add to the network.
 
     """
+
+    split_edge = network_split_edge
+    join_edges = network_join_edges
+    smooth = network_smooth_centroid
+
+    is_crossed = network_is_crossed
+    is_planar = network_is_planar
+    is_planar_embedding = network_is_planar_embedding
+    is_xy = network_is_xy
+    count_crossings = network_count_crossings
+    find_crossings = network_find_crossings
+    embed_in_plane = network_embed_in_plane
+
+    find_cycles = network_find_cycles
 
     def __init__(self, default_node_attributes=None, default_edge_attributes=None, **kwargs):
         _default_node_attributes = {"x": 0.0, "y": 0.0, "z": 0.0}
@@ -355,6 +385,28 @@ class Network(Graph):
     # accessors
     # --------------------------------------------------------------------------
 
+    def shortest_path(self, u, v):
+        """Find the shortest path between two nodes using the A* algorithm.
+
+        Parameters
+        ----------
+        u : hashable
+            The identifier of the start node.
+        v : hashable
+            The identifier of the end node.
+
+        Returns
+        -------
+        list[hashable] | None
+            The path from root to goal, or None, if no path exists between the vertices.
+
+        See Also
+        --------
+        :meth:`compas.topology.astar_shortest_path`
+
+        """
+        return astar_shortest_path(self.adjacency, u, v)
+
     # --------------------------------------------------------------------------
     # node attributes
     # --------------------------------------------------------------------------
@@ -663,50 +715,24 @@ class Network(Graph):
         a, b = self.edge_coordinates(edge)
         return distance_point_point(a, b)
 
+    # --------------------------------------------------------------------------
+    # transformations
+    # --------------------------------------------------------------------------
 
-# =============================================================================
-# Add additional methods
-# =============================================================================
+    def transform(self, transformation):
+        """Transform all nodes of the network.
 
-from .operations.split import network_split_edge  # noqa: E402
-from .combinatorics import network_is_connected  # noqa: E402
-from .complementarity import network_complement  # noqa: E402
-from .duality import network_find_cycles  # noqa: E402
-from .transformations import network_transform  # noqa: E402
-from .transformations import network_transformed  # noqa: E402
-from .traversal import network_shortest_path  # noqa: E402
-from .smoothing import network_smooth_centroid  # noqa: E402
-from .planarity import network_count_crossings  # noqa: E402
-from .planarity import network_find_crossings  # noqa: E402
-from .planarity import network_is_crossed  # noqa: E402
-from .planarity import network_is_xy  # noqa: E402
+        Parameters
+        ----------
+        transformation : :class:`Transformation`
+            The transformation used to transform the nodes.
 
-Network.complement = network_complement  # type: ignore
-Network.count_crossings = network_count_crossings  # type: ignore
-Network.find_crossings = network_find_crossings  # type: ignore
-Network.find_cycles = network_find_cycles  # type: ignore
-Network.is_connected = network_is_connected  # type: ignore
-Network.is_crossed = network_is_crossed  # type: ignore
-Network.is_xy = network_is_xy  # type: ignore
-Network.shortest_path = network_shortest_path  # type: ignore
-Network.smooth = network_smooth_centroid  # type: ignore
-Network.split_edge = network_split_edge  # type: ignore
-Network.transform = network_transform  # type: ignore
-Network.transformed = network_transformed  # type: ignore
+        Returns
+        -------
+        None
 
-if not compas.IPY:
-    from .matrices import network_adjacency_matrix
-    from .matrices import network_connectivity_matrix
-    from .matrices import network_degree_matrix
-    from .matrices import network_laplacian_matrix
-    from .planarity import network_embed_in_plane
-    from .planarity import network_is_planar
-    from .planarity import network_is_planar_embedding
-
-    Network.adjacency_matrix = network_adjacency_matrix  # type: ignore
-    Network.connectivity_matrix = network_connectivity_matrix  # type: ignore
-    Network.degree_matrix = network_degree_matrix  # type: ignore
-    Network.embed_in_plane = network_embed_in_plane  # type: ignore
-    Network.is_planar = network_is_planar  # type: ignore
-    Network.is_planar_embedding = network_is_planar_embedding  # type: ignore
-    Network.laplacian_matrix = network_laplacian_matrix  # type: ignore
+        """
+        nodes = self.nodes_attributes("xyz")
+        points = transform_points(nodes, transformation)
+        for point, node in zip(points, self.nodes()):
+            self.node_attributes(node, "xyz", point)